sql/row: move the base FK to a separate file, add docs
Release note: None
